// src/models/IoK8sApiAutoscalingV2HorizontalPodAutoscalerBehavior/types.ts
import { IoK8sApiAutoscalingV2HPAScalingRulesModelOptions as IoK8sApiAutoscalingV2HPAScalingRules } from '../IoK8sApiAutoscalingV2HPAScalingRules';

export interface IoK8sApiAutoscalingV2HorizontalPodAutoscalerBehaviorModelOptions {
  scaleDown?: IoK8sApiAutoscalingV2HPAScalingRules;
  scaleUp?: IoK8sApiAutoscalingV2HPAScalingRules;
}